+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - `Cancel` not canceling middleware stacks.
+- `404 Not Found` not being sent when no handler is found.
 
 ## [0.5.0] - 2017-12-21
 ### Added